+++ include/kanji_tools/quiz/QuizLauncher.h
@@ -32,7 +32,7 @@ public:
   // or Group based). 'quizType' can be 'f', 'g', 'k', 'l', 'm' or 'p' for the
   // type of quiz/review and 'questionList' can also be provided (values depend
   // on quiz type - see Quiz.cpp 'HelpMessage' for details).
-  void start(OptChar quizType, OptChar questionList, int question = 0,
+  void start(OptChar quizType, OptChar questionList, size_t question = 0,
              bool showMeanings = false);
 
   [[nodiscard]] auto& log(bool heading = false) const {
@@ -61,15 +61,15 @@ private:
 
   [[nodiscard]] const Data& data() const { return _groupData->data(); }
 
-  void startListQuiz(int question, bool showMeanings, KanjiInfo excludeField,
+  void startListQuiz(size_t question, bool showMeanings, KanjiInfo excludeField,
                      const List&) const;
-  void startGroupQuiz(int question, bool showMeanings, OptChar questionList,
+  void startGroupQuiz(size_t question, bool showMeanings, OptChar questionList,
                       const GroupData::List& list) const;
 
   // 'processProgramModeArg' is called for '-r' and '-t' args and sets
   // '_programMode'. It can also set '_questionOrder' depending on the value of
   // 'arg' and returns the question to start from.
-  [[nodiscard]] int processProgramModeArg(const std::string& arg);
+  [[nodiscard]] size_t processProgramModeArg(const std::string& arg);
 
   // 'processKanjiArg' is called when a kanji arg is passed to the program (see
   // 'HelpMessage' in QuizLauncher.cpp)
